исправлены замечания

+++ app/Http/Controllers/Crm/CompanyController.php
@@ -3,29 +3,19 @@
 namespace App\Http\Controllers\Crm;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\PageRequest;
 use App\Http\Traits\TableHelper;
 use App\Models\Company;
 use App\Http\Requests\Crm\Company\StoreCompanyRequest;
 use App\Http\Requests\Crm\Company\UpdateCompanyRequest;
 use App\View\Components\AdminButton;
 use App\View\Components\ModalInfo;
-use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Redirect;
 use Illuminate\Support\Facades\Storage;
 
 class CompanyController extends Controller
 {
     use TableHelper;
-
-    /**
-     * Instantiate a new controller instance.
-     *
-     * @return void
-     */
-    public function __construct()
-    {
-        $this->middleware(['auth', 'verified', 'set_language']);
-    }
 
     /**
      * Display a listing of the resource.
@@ -39,30 +29,29 @@ class CompanyController extends Controller
 
     /**
      * AJAX pagination.
-     * @param Request $request
+     *
+     * @param PageRequest $request
+     * @return \Illuminate\Http\JsonResponse
      */
-    public function page(Request $request)
+    public function page(PageRequest $request)
     {
-        /** @var string $draw */
-        /** @var string $start */
-        /** @var string $rowsPerPage */
-        /** @var string $columnIndex */
-        /** @var string $columnName */
-        /** @var string $columnSortOrder */
-        /** @var string $searchValue */
-        extract($this->getPageParams($request));
+        $page = (int) $request->start > 1 ? ($request->start / $request->length) + 1 : 1;
+        $columnIndex = $request->order[0]['column']; // Column index
+        $columnName = $request->columns[$columnIndex]['data']; // Column name
+        $columnSortOrder = $request->order[0]['dir'] ?? 'asc'; // asc or desc
+        $searchValue = $request->search['value']; // Search value
 
         $lang = config('app.locale');
         if ($columnName === 'name')
-            $columnName .= "_$lang";
+            $columnName = "cast($columnName->>'$.$lang' as char)";
 
         // Fetch records
-        $records = Company::orderBy($columnName, $columnSortOrder)
-            ->where('name_en', 'like', '%' .$searchValue . '%')
-            ->orWhere('name_ua', 'like', '%' .$searchValue . '%')
-            ->skip($start)
-            ->take($rowsPerPage)
-            ->get();
+        $records = Company::orderByRaw("$columnName $columnSortOrder")
+            ->when($searchValue, fn ($q) => $q
+                ->where('name->en', 'like', '%' .$searchValue . '%')
+                ->orWhere('name->ua', 'like', '%' .$searchValue . '%')
+            )
+            ->paginate($request->length, ['*'], 'page', $page);
 
         $data = [];
         foreach($records as $record)
@@ -71,7 +60,7 @@ class CompanyController extends Controller
             $deleteButton = new AdminButton('companies', 'company', 'destroy', $record['id'], true);
             $data[] = [
                 "id"            => $record['id'],
-                "company_name"  => $record["name_$lang"],
+                "name"          => $record['name'],
                 "email"         => $record['email'],
                 "phone"         => $record['phone'],
                 "website"       => '<a href="'.$record['website'].'" target="_blank">'.$record['website'].'</a>',
@@ -83,26 +72,27 @@ class CompanyController extends Controller
         // Total records
         $totalRecords = Company::select('count(*) as allcount')->count();
         $totalFilteredRecords = Company::select('count(*) as allcount')
-            ->where('name_en', 'like', '%' .$searchValue . '%')
-            ->orWhere('name_ua', 'like', '%' .$searchValue . '%')
+            ->when($searchValue, fn ($q) => $q
+                ->where('name->en', 'like', '%' .$searchValue . '%')
+                ->orWhere('name->ua', 'like', '%' .$searchValue . '%')
+            )
             ->count();
 
-        echo json_encode([
-            "draw" => intval($draw),
+        return response()->json([
+            "draw" => intval($request->draw),
             "recordsTotal" => $totalRecords,
             "recordsFiltered" => $totalFilteredRecords,
             "data" => $data
-        ], JSON_UNESCAPED_UNICODE);
+        ]);
     }
 
     /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Contracts\View\View
+     * Get form controls.
+     * @return array[]
      */
-    public function create()
+    public function controls()
     {
-        $controls = [
+        return [
             [
                 'type' => 'text',
                 'name' => 'name_ua',
@@ -135,7 +125,20 @@ class CompanyController extends Controller
                 'value' => '/img/logo.png'
             ]
         ];
-        return view('admin.company.create', ['controls' => $controls]);
+    }
+
+    /**
+     * Show the form for creating a new resource.
+     *
+     * @return \Illuminate\Contracts\View\View
+     */
+    public function create()
+    {
+        $action = 'add';
+        $route = route('companies.store');
+        $method = 'post';
+        $controls = $this->controls();
+        return view('admin.company.edit', compact('action', 'route', 'method', 'controls'));
     }
 
     /**
@@ -147,24 +150,21 @@ class CompanyController extends Controller
     public function store(StoreCompanyRequest $request)
     {
         // save entry
-        $company = Company::create([
-            'name_en' => $request['name_en'],
-            'name_ua' => $request['name_ua'],
-            'email'   => $request['email'],
-            'phone'   => $request['phone'],
-            'website'   => $request['website']
-        ]);
-
+        $path = '';
         if (isset($request['logo']))
-        {
-            $ext = $request['logo']->getClientOriginalExtension();
-            $company->logo = Storage::disk('public')->putFileAs(
-                'companies',
-                $request['logo'],
-                "logo$company->id.$ext"
-            );
-            $company->save();
-        }
+            $path = $request->file('logo')->store('companies');
+
+        $company = Company::create([
+            'name'      => [
+                'en'    => $request['name_en'],
+                'ua'    => $request['name_ua'],
+            ],
+            'email'     => $request['email'],
+            'phone'     => $request['phone'],
+            'website'   => $request['website'],
+            'logo'      => $path
+        ]);
+        $company->save();
         // show result
         $genitive = $this->mb_ucfirst($this->getGenitiveCaseName('App\Models\Company'));
         $modal = new ModalInfo(
@@ -177,73 +177,64 @@ class CompanyController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Models\Company  $company
-     * @return \Illuminate\Contracts\View\View
+     * @param Company $company
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|\Illuminate\Http\JsonResponse
      */
     public function edit(Company $company)
     {
-        $controls = [
-            [
-                'type' => 'text',
-                'name' => 'name_ua',
-                'placeholder' => __('admin.company_name_ua'),
-                'value' => $company['name_ua']
-            ],
-            [
-                'type' => 'text',
-                'name' => 'name_en',
-                'placeholder' => __('admin.company_name_en'),
-                'value' => $company['name_en']
-            ],
-            [
-                'type' => 'email',
-                'name' => 'email',
-                'placeholder' => __('admin.email'),
-                'value' => $company['email']
-            ],
-            [
-                'type' => 'text',
-                'name' => 'phone',
-                'placeholder' => __('admin.phone'),
-                'value' => $company['phone']
-            ],
-            [
-                'type' => 'text',
-                'name' => 'website',
-                'placeholder' => __('admin.website'),
-                'value' => $company['website']
-            ],
-            [
-                'type' => 'image',
-                'name' => 'logo',
-                'placeholder' => __('admin.company_logo'),
-                'value' => empty($company['logo']) ? '/img/logo.png' : '/storage/'.$company['logo']
-            ]
-        ];
-        return view('admin.company.edit', ['controls' => $controls, 'id' => $company['id']]);
+        if (auth()->user()->id !== $company->created_by()->first()->id)
+            abort(403);
+
+        $action = 'edit';
+        $route = route('companies.update', ['company' => $company['id']]);
+        $method = 'put';
+        $controls = $this->controls();
+        foreach ($controls as &$control)
+        {
+            $field = $control['name'];
+            if ($field === 'logo')
+            {
+                empty($company['logo']) ?: $control['value'] = '/storage/'.$company['logo'];
+            }
+            else if (str_contains($field, 'name'))
+            {
+                $lang = explode('_', $field)[1];
+                $control['value'] = $company->getTranslation('name', $lang);
+            }
+            else
+            {
+                $control['value'] = $company[$field];
+            }
+        }
+        return view('admin.company.edit', compact('action', 'route', 'method', 'controls'));
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  \App\Http\Requests\Crm\Company\UpdateCompanyRequest  $request
-     * @param  \App\Models\Company  $company
-     * @return \Illuminate\Http\RedirectResponse
+     * @param UpdateCompanyRequest $request
+     * @param Company $company
+     * @return \Illuminate\Http\JsonResponse|\Illuminate\Http\RedirectResponse
      */
     public function update(UpdateCompanyRequest $request, Company $company)
     {
-        $company->name_ua = $request['name_ua'];
-        $company->name_en = $request['name_en'];
+        if (auth()->user()->id !== $company->created_by()->first()->id)
+            abort(403);
+
+        $company->setTranslation('name', 'ua', $request['name_ua']);
+        $company->setTranslation('name', 'en', $request['name_en']);
         $company->email = $request['email'];
         $company->phone = $request['phone'];
         $company->website = $request['website'];
         if (isset($request['logo']))
         {
-            // delete the file in case of a different extension
-            Storage::disk('public')->delete($company->logo);
+            // save old file path
+            $oldLogo = $company->logo;
             // save new file
-            $ext = $request['logo']->getClientOriginalExtension();
-            $company->logo = Storage::disk('public')->putFileAs('companies', $request['logo'], "logo$company->id.$ext");
+            $company->logo = $request->file('logo')->store('companies');
+            // delete the old file
+            if (!empty($oldLogo))
+                Storage::delete($oldLogo);
         }
         $company->save();
         // show result
@@ -258,17 +249,22 @@ class CompanyController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\Company  $company
-     * @return \Illuminate\Http\RedirectResponse
+     * @param Company $company
+     * @return \Illuminate\Http\JsonResponse|\Illuminate\Http\RedirectResponse
      */
     public function destroy(Company $company)
     {
+        if (auth()->user()->id !== $company->created_by()->first()->id)
+            abort(403);
+
         $genitive = $this->mb_ucfirst($this->getGenitiveCaseName('App\Models\Company'));
         $modal = new ModalInfo(
             __('admin.delete_success_header', ['entity' => $genitive]),
             __('admin.delete_success_text', ['entity' => $genitive, 'id' => $company['id']]),
         );
 
+        if (!empty($company->logo))
+            Storage::delete($company->logo);
         $company->delete();
 
         return Redirect::route('companies.index')->with('modal', $modal);
